refactor: move editor stories into Features directory
