+++ packages/serverless-workflow-diagram-editor/stories/features/DiagramEditor.stories.ts
@@ -84,7 +84,7 @@ timeout:
 
 const meta = {
   id: "diagram-editor",
-  title: "Example/Diagram-Editor",
+  title: "Features/Diagram-Editor",
   component: DiagramEditor,
   // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/writing-docs/autodocs
   tags: ["autodocs"],
