+++ packages/serverless-workflow-diagram-editor/stories/features/DiagramEditorDragNDrop.stories.ts
@@ -14,21 +14,27 @@
  * limitations under the License.
  */
 
-import {
-  DiagramEditor as Component,
-  DiagramEditorProps,
-} from "../src/diagram-editor/DiagramEditor";
+import type { Meta, StoryObj } from "@storybook/react-vite";
 
-/** Primary UI component for user interaction */
-export const DiagramEditor = ({ ...props }: DiagramEditorProps) => {
-  return (
-    <div style={{ height: "100vh" }}>
-      <Component
-        content={props.content}
-        isReadOnly={props.isReadOnly}
-        locale={props.locale}
-        colorMode={props.colorMode}
-      />
-    </div>
-  );
+import { DiagramEditorDragNDrop } from "./DiagramEditorDragNDrop";
+
+const meta = {
+  id: "diagram-editor-drag-n-drop",
+  title: "Features/Diagram-Editor-Drag-N-Drop",
+  component: DiagramEditorDragNDrop,
+  tags: ["autodocs"],
+  parameters: {
+    layout: "fullscreen",
+  },
+  args: {},
+} satisfies Meta<typeof DiagramEditorDragNDrop>;
+
+export default meta;
+type Story = StoryObj<typeof meta>;
+
+export const Component: Story = {
+  args: {
+    isReadOnly: true,
+    locale: "en",
+  },
 };
